+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'everyteen-be'
+rootProject.name = 'moongch-be'
