+++ src/main/java/com/devsco/moongch/MoongchApplication.java
@@ -1,11 +1,11 @@
-package com.devsco.everyteen;
+package com.devsco.moongch;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class EveryteenApplication {
+public class MoongchApplication {
   public static void main(String[] args) {
-    SpringApplication.run(EveryteenApplication.class, args);
+    SpringApplication.run(MoongchApplication.class, args);
   }
 }
