+++ src/main/java/com/devsco/moongch/exception/ExceptionHandlerAdvice.java
@@ -1,4 +1,4 @@
-package com.devsco.everyteen.exception;
+package com.devsco.moongch.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ProblemDetail;
