+++ src/test/java/com/devsco/moongch/MoongchApplicationTests.java
@@ -1,10 +1,10 @@
-package com.devsco.everyteen;
+package com.devsco.moongch;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class EveryteenApplicationTests {
+class MoongchApplicationTests {
 
   @Test
   void contextLoads() {
