+++ src/test/java/com/devsco/moongch/exception/MockController.java
@@ -1,4 +1,4 @@
-package com.devsco.everyteen.exception;
+package com.devsco.moongch.exception;
 
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotNull;
